Rename namespace `DayNightCycle` -> `Cycle`
