+++ Cycle/Weather/WeatherCycle.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using Godot;
-using untitledplantgame.DayNightCycle.UI;
+using untitledplantgame.Cycle.UI;
 
-namespace untitledplantgame.DayNightCycle;
+namespace untitledplantgame.Cycle;
 
 public enum Weather
 {
